Updated test code per review feedback
